libs/python/dp-core/pyproject.toml
[project]
name = "dp-core"
version = "0.1.0"
description = "Core utilities for the data platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Data Platform Team" }]

dependencies = [
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "structlog>=24.1",
    "tenacity>=8.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dp_core"]
